already pass tested in real hardware

+++ init.sh
@@ -11,16 +11,15 @@ configgit() {
     git config --global user.email nosde@protonmail.com
 }
 
-installnvcode() {
-    echo 'Installing NVcode'
-    sudo npm install -g tree-sitter-cli pyright bash-language-server prettier
-    luarocks install --server=https://luarocks.org/dev luaformatter
+installenev() {
+    echo 'Installing ENEv'
+    sudo npm install -g tree-sitter-cli pyright bash-language-server prettier --force
+    sudo luarocks install --server=https://luarocks.org/dev luaformatter
     bash <(curl -s https://raw.githubusercontent.com/Untesler/LunarVim/master/utils/installer/install.sh)
-    # Patch my own config
-    cp -rf $dotpath'/patch_nvcode/*' $HOME'/.config/nvim/'
 }
 
 initsshgithub() {
+    # TODO: Fix this func (sshpath not found in ssh-add)
     echo 'Initializing SSH for Github'
     echo -n 'Enter email : '
     read email
@@ -60,9 +59,15 @@ installpolybarthemes() {
 }
 
 installanaconda() {
+    # TODO: Seperate this func into another script because need to restart the terminal to take affect after installing conda
     echo 'Installing Anaconda3'
     if [ -x "$(command -v conda)" ]; then
       echo 'Anaconda is already installed!'
+      sudo conda update --yes conda
+      sudo conda update --yes anaconda
+      source ~/.bash_profile
+      # Install necessary packages
+      pip install neovim-remote pynvim ueberzug sphinx meson
     else
       # Scrape the latest Anaconda
       anaconda_url=$(wget -O - https://www.anaconda.com/distribution/ 2>/dev/null | sed -ne 's@.*\(https:\/\/repo\.anaconda\.com\/archive\/Anaconda3-.*-Linux-x86_64\.sh\)\">64-Bit (x86) Installer.*@\1@p')
@@ -72,13 +77,9 @@ installanaconda() {
       sudo sh anaconda_script.sh
       rm anaconda_script.sh
     fi
-    conda update --yes conda
-    conda update --yes anaconda
-    source ~/.bash_profile
-    # Install necessary packages
-    pip install neovim-remote pynvim ueberzug sphinx meson
 }
 
+sudo pwd
 echo 'Initializing System'
 
 answer='n'
@@ -91,30 +92,35 @@ while [ "$answer" == "n" ]; do
 done
 
 # If not have these folders, then create 
-[ ! -d "$HOME/Scripts" ] && mkdir ~/Scripts
-[ ! -d "$HOME/Downloads/Torrents" ] && mkdir ~/Downloads/Torrents
-[ ! -d "$HOME/Pictures/Wallpapers" ] && mkdir ~/Pictures/Wallpapers
-[ ! -d "$HOME/Pictures/Avatars" ] && mkdir ~/Pictures/Avatars
-[ ! -d "$HOME/Application" ] && mkdir ~/Application
-[ ! -d "$HOME/Repos" ] && mkdir ~/Repos
+[ ! -d "$HOME/Scripts" ] && mkdir $HOME'/Scripts'
+[ ! -d "$HOME/Downloads/Torrents" ] && mkdir -p  $HOME'/Downloads/Torrents'
+[ ! -d "$HOME/Pictures/Wallpapers" ] && mkdir -p $HOME'/Pictures/Wallpapers'
+[ ! -d "$HOME/Pictures/Avatars" ] && mkdir -p $HOME'/Pictures/Avatars'
+[ ! -d "$HOME/Application" ] && mkdir $HOME'/Application'
+[ ! -d "$HOME/Repos" ] && mkdir $HOME'/Repos'
 
 sudo pacman -Syu
 # rsm is for runit service manager
-echo -n "Would you like to install rsm(runit services manager) now (y/n)? "
-read answer
-[ "$answer" != "${answer#[Yy]}" ] && sudo pacman -S rsm
+if ! command -v rsm &> /dev/null; then
+	echo -n "Would you like to install rsm(runit services manager) now (y/n)? "
+	read answer
+	[ "$answer" != "${answer#[Yy]}" ] && sudo pacman -S rsm --needed --noconfirm
+fi
 
 echo 'Installing packages'
 # maybe need to add more in order to build a picom with mesa
-sudo pacman -S base-devel pkg-config boost git feh xautolock catdoc pandoc cmake wget fish fzf go rust lua jre-openjdk jdk-oprenjdk jq neofetch vim nodejs npm luarocks meson nvidia-dkms nvidia-settings python-sphinx ranger rofi alsa tree-sitter ueberzug unoconv xclip xsel zathura zathura-cb zathura-djvu zathura-pdf-mupdf ripgrep uthash mesa check fd firefox libev xcb-util libxcb libconfig dbus dunst keepassxc ufw rtorrent exa bat kitty rsync dragon mediainfo tree sxiv --needed --noconfirm
+sudo pacman -S base-devel pkg-config boost git feh xautolock catdoc pandoc cmake wget fish fzf go rust lua jre-openjdk jdk-openjdk jq neofetch vim nodejs npm luarocks meson nvidia-dkms nvidia-settings python-sphinx ranger rofi alsa tree-sitter unoconv xclip xsel zathura zathura-cb zathura-djvu zathura-pdf-mupdf ripgrep uthash mesa check fd firefox libev xcb-util libxcb libconfig dbus dunst keepassxc ufw rtorrent exa bat kitty rsync dragon mediainfo tree sxiv --needed --noconfirm
 configgit
 
 echo 'Installing Paru for managing AUR packages'
-git clone https://aur.archlinux.org/paru.git
-cd paru && sudo makepkg -si
-cd .. && sudo rm -rf paru
+if ! command -v paru &> /dev/null; then
+	git clone https://aur.archlinux.org/paru.git
+	cd paru && makepkg -si
+	cd .. && sudo rm -rf paru
+fi
 echo 'Installing packages from AUR'
-paru -S alacritty audacity discord spotify obs-studio flameshot nvm neovim-nightly-git moc-pulse-svn python-ueberzug-git polybar-git lazygit gnome-disk-utility microsoft-edge-dev-bin mpc-qt-git neovide-git spotifyd spotifyd-runit spotify-dev spotify-tui drive-git i3lock-color translate-shell zoxide-bin nnn-nerd python-pynvim --needed --noconfirm
+# drive-git
+paru -S alacritty audacity discord spotify obs-studio flameshot nvm neovim-nightly-git moc-pulse-svn python-ueberzug-git polybar-git lazygit gnome-disk-utility microsoft-edge-dev-bin mpc-qt-git neovide-git spotifyd spotifyd-runit spotify-tui-bin i3lock-color translate-shell zoxide-bin nnn-nerd python-pynvim --needed
 
 cp -rf $dotpath'/bspwm' $HOME'/.config'
 cp -rf $dotpath'/ranger' $HOME'/.config'
@@ -126,7 +132,25 @@ cp -rf $dotpath'/.moc' $HOME
 cp $dotpath/.tmux.conf $HOME
 cp $dotpath/.rtorrent.rc $HOME
 cp $dotpath/.xinitrc $HOME
-sh $dotpath'/fonts/install.sh'
+echo 'Installing fonts'
+
+paru -Sy ttf-iosevka --needed
+
+sudo chmod 555 $dotpath'/fonts/OTF/'
+sudo chmod 555 $dotpath'/fonts/TTF/'
+
+sudo cp -rf $dotpath'/fonts/OTF/' /usr/share/fonts
+sudo cp -rf $dotpath'/fonts/TTF/' /usr/share/fonts
+
+[ ! -d "$HOME/.local" ] && mkdir $HOME/.local
+[ ! -d "$HOME/.local/share" ] && mkdir $HOME/.local/share
+[ ! -d "$HOME/.local/share/fonts" ] && mkdir $HOME/.local/share/fonts
+
+sudo cp -rf $dotpath'/fonts/OTF/' $HOME/.local/share/fonts
+sudo cp -rf $dotpath'/fonts/TTF/' $HOME/.local/share/fonts
+
+sudo fc-cache
+
 echo -n "Would you like to swap CapsLock key with Grave(~,\`) key (y/n)? "
 read answer
 [ "$answer" != "${answer#[Yy]}" ] && cp $dotpath'/.Xmodmap' $HOME
@@ -145,7 +169,7 @@ cat $dotpath | \
     while read package; do
         fish -c 'fisher install '$package
     done
-cp $dotpath'/fish/config.fish' $HOME'.config/fish'
+cp $dotpath'/fish/config.fish' $HOME'/.config/fish'
 
 echo 'Installing Polybar-Theme'
 installpolybarthemes
@@ -157,6 +181,6 @@ echo -n "Would you like to initialize ssh for github now (y/n)? "
 read answer
 [ "$answer" != "${answer#[Yy]}" ] && initsshgithub
 
-echo -n "Would you like to install NVcode now (y/n)? "
+echo -n "Would you like to install ENEv now (y/n)? "
 read answer
-[ "$answer" != "${answer#[Yy]}" ] && installnvcode
+[ "$answer" != "${answer#[Yy]}" ] && installenev
